requirements.txt
python-telegram-bot==13.7
spotipy==2.23.0
moviepy==1.0.3
Pillow==10.0.0
numpy==1.24.3
requests==2.31.0
python-dotenv==1.0.0
urllib3==1.26.15